+++ api/v1/views/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/python3
+
+from flask import Blueprint
+
+app_views = Blueprint('app_views', __name__, url_prefix='/api/v1')
+from api.v1.views.index import *
+from api.v1.views.states import *
